# schema/openapi/backend.json
{
 "components": {
  "schemas": {
   "Body_upload_api_books_post": {
    "properties": {
     "file": {
      "contentMediaType": "application/octet-stream",
      "title": "File",
      "type": "string"
     }
    },
    "required": [
     "file"
    ],
    "title": "Body_upload_api_books_post",
    "type": "object"
   },
   "BookRow": {
    "properties": {
     "name": {
      "title": "Name",
      "type": "string"
     },
     "runs": {
      "items": {
       "$ref": "#/components/schemas/RunRow"
      },
      "title": "Runs",
      "type": "array"
     }
    },
    "required": [
     "name",
     "runs"
    ],
    "title": "BookRow",
    "type": "object"
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "Results": {
    "properties": {
     "records": {
      "items": {
       "$ref": "#/components/schemas/SeriesRow"
      },
      "title": "Records",
      "type": "array"
     },
     "when": {
      "title": "When",
      "type": "number"
     }
    },
    "required": [
     "when",
     "records"
    ],
    "title": "Results",
    "type": "object"
   },
...
   "RunRow": {
    "properties": {
     "complete": {
      "title": "Complete",
      "type": "boolean"
     },
     "identity": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Identity"
     },
     "kind": {
      "title": "Kind",
      "type": "string"
     },
     "label": {
      "title": "Label",
      "type": "string"
     },
     "level": {
      "title": "Level",
      "type": "string"
     },
     "pages": {
      "title": "Pages",
      "type": "integer"
     },
     "when": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "When"
     }
    },
    "required": [
     "kind",
     "label",
     "identity",
     "pages",
     "complete",
     "level",
     "when"
    ],
    "title": "RunRow",
    "type": "object"
   },
...
  }
 },
 "info": {
  "title": "backend",
  "version": "0.1.0"
 },
 "openapi": "3.1.0",
 "paths": {
  "/api/books": {
   "get": {
    "operationId": "books_api_books_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/BookRow"
         },
         "title": "Response Books Api Books Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "summary": "Books"
   },
   "post": {
    "operationId": "upload_api_books_post",
    "parameters": [
     {
      "in": "query",
      "name": "name",
      "required": false,
      "schema": {
       "default": "",
       "title": "Name",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_upload_api_books_post"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Upload Api Books Post",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Upload"
   }
  },
  "/api/books/{root}/{name}": {
   "delete": {
    "operationId": "delete_book_api_books__root___name__delete",
    "parameters": [
     {
      "in": "path",
      "name": "root",
      "required": true,
      "schema": {
       "title": "Root",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "name",
      "required": true,
      "schema": {
       "title": "Name",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Delete Book Api Books  Root   Name  Delete",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Delete Book"
   }
  },
  "/api/books/{root}/{name}/pages/{index}/image": {
   "get": {
    "operationId": "page_image_api_books__root___name__pages__index__image_get",
    "parameters": [
     {
      "in": "path",
      "name": "root",
      "required": true,
      "schema": {
       "title": "Root",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "name",
      "required": true,
      "schema": {
       "title": "Name",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "index",
      "required": true,
      "schema": {
       "title": "Index",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "dpi",
      "required": false,
      "schema": {
       "default": 110.0,
       "title": "Dpi",
       "type": "number"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Page Image"
   }
  },
  "/api/books/{root}/{name}/runs": {
   "get": {
    "operationId": "runs_api_books__root___name__runs_get",
    "parameters": [
     {
      "in": "path",
      "name": "root",
      "required": true,
      "schema": {
       "title": "Root",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "name",
      "required": true,
      "schema": {
       "title": "Name",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/RunRow"
         },
         "title": "Response Runs Api Books  Root   Name  Runs Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Runs"
   }
  },
  "/api/books/{root}/{name}/runs/{kind}/{label}": {
   "get": {
    "operationId": "run_api_books__root___name__runs__kind___label__get",
    "parameters": [
     {
      "in": "path",
      "name": "root",
      "required": true,
      "schema": {
       "title": "Root",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "name",
      "required": true,
      "schema": {
       "title": "Name",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "kind",
      "required": true,
      "schema": {
       "title": "Kind",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "label",
      "required": true,
      "schema": {
       "title": "Label",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/RunInfo"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Run"
   }
  },
  "/api/books/{root}/{name}/runs/{kind}/{label}/crops/{anchor}": {
   "get": {
    "operationId": "crop_api_books__root___name__runs__kind___label__crops__anchor__get",
    "parameters": [
     {
      "in": "path",
      "name": "root",
      "required": true,
      "schema": {
       "title": "Root",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "name",
      "required": true,
      "schema": {
       "title": "Name",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "kind",
      "required": true,
      "schema": {
       "title": "Kind",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "label",
      "required": true,
      "schema": {
       "title": "Label",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "anchor",
      "required": true,
      "schema": {
       "title": "Anchor",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Crop"
   }
  },
  "/api/books/{root}/{name}/runs/{kind}/{label}/document": {
   "get": {
    "operationId": "document_api_books__root___name__runs__kind___label__document_get",
    "parameters": [
     {
      "in": "path",
      "name": "root",
      "required": true,
      "schema": {
       "title": "Root",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "name",
      "required": true,
      "schema": {
       "title": "Name",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "kind",
      "required": true,
      "schema": {
       "title": "Kind",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "label",
      "required": true,
      "schema": {
       "title": "Label",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Document Api Books  Root   Name  Runs  Kind   Label  Document Get",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Document"
   }
  },
  "/api/books/{root}/{name}/runs/{kind}/{label}/pages/{index}": {
   "get": {
    "operationId": "page_api_books__root___name__runs__kind___label__pages__index__get",
    "parameters": [
     {
      "in": "path",
      "name": "root",
      "required": true,
      "schema": {
       "title": "Root",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "name",
      "required": true,
      "schema": {
       "title": "Name",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "kind",
      "required": true,
      "schema": {
       "title": "Kind",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "label",
      "required": true,
      "schema": {
       "title": "Label",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "index",
      "required": true,
      "schema": {
       "title": "Index",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Page Api Books  Root   Name  Runs  Kind   Label  Pages  Index  Get",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Page"
   }
  },
...
  "/api/books/{root}/{name}/runs/{kind}/{label}/pages/{index}/pairs": {
   "get": {
    "operationId": "pairs_api_books__root___name__runs__kind___label__pages__index__pairs_get",
    "parameters": [
     {
      "in": "path",
      "name": "root",
      "required": true,
      "schema": {
       "title": "Root",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "name",
      "required": true,
      "schema": {
       "title": "Name",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "kind",
      "required": true,
      "schema": {
       "title": "Kind",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "label",
      "required": true,
      "schema": {
       "title": "Label",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "index",
      "required": true,
      "schema": {
       "title": "Index",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Pairs Api Books  Root   Name  Runs  Kind   Label  Pages  Index  Pairs Get",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Pairs"
   }
  },
  "/api/books/{root}/{name}/runs/{kind}/{label}/results": {
   "get": {
    "operationId": "results_api_books__root___name__runs__kind___label__results_get",
    "parameters": [
     {
      "in": "path",
      "name": "root",
      "required": true,
      "schema": {
       "title": "Root",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "name",
      "required": true,
      "schema": {
       "title": "Name",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "kind",
      "required": true,
      "schema": {
       "title": "Kind",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "label",
      "required": true,
      "schema": {
       "title": "Label",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Results"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Results"
   }
  },
  "/api/books/{root}/{name}/runs/{kind}/{label}/series": {
   "get": {
    "operationId": "series_api_books__root___name__runs__kind___label__series_get",
    "parameters": [
     {
      "in": "path",
      "name": "root",
      "required": true,
      "schema": {
       "title": "Root",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "name",
      "required": true,
      "schema": {
       "title": "Name",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "kind",
      "required": true,
      "schema": {
       "title": "Kind",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "label",
      "required": true,
      "schema": {
       "title": "Label",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/SeriesRow"
         },
         "title": "Response Series Api Books  Root   Name  Runs  Kind   Label  Series Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Series"
   }
  },
  "/api/books/{root}/{name}/truth": {
   "post": {
    "operationId": "start_truth_api_books__root___name__truth_post",
    "parameters": [
     {
      "in": "path",
      "name": "root",
      "required": true,
      "schema": {
       "title": "Root",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "name",
      "required": true,
      "schema": {
       "title": "Name",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TruthStart"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Start Truth"
   }
  },
  "/api/books/{root}/{name}/truth/pages/{index}": {
   "get": {
    "operationId": "truth_page_api_books__root___name__truth_pages__index__get",
    "parameters": [
     {
      "in": "path",
      "name": "root",
      "required": true,
      "schema": {
       "title": "Root",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "name",
      "required": true,
      "schema": {
       "title": "Name",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "index",
      "required": true,
      "schema": {
       "title": "Index",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Truth Page Api Books  Root   Name  Truth Pages  Index  Get",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Truth Page"
   },
   "put": {
    "operationId": "put_truth_page_api_books__root___name__truth_pages__index__put",
    "parameters": [
     {
      "in": "path",
      "name": "root",
      "required": true,
      "schema": {
       "title": "Root",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "name",
      "required": true,
      "schema": {
       "title": "Name",
       "type": "string"
      }
     },
     {
      "in": "path",
      "name": "index",
      "required": true,
      "schema": {
       "title": "Index",
       "type": "integer"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "additionalProperties": true,
        "title": "Body",
        "type": "object"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Layer"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Put Truth Page"
   }
  },
...
  "/api/fleet/placements/{pid}": {
   "delete": {
    "operationId": "fleet_stop_api_fleet_placements__pid__delete",
    "parameters": [
     {
      "in": "path",
      "name": "pid",
      "required": true,
      "schema": {
       "title": "Pid",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Fleet Stop Api Fleet Placements  Pid  Delete",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Fleet Stop"
   }
  },
  "/api/jobs": {
   "get": {
    "operationId": "jobs_api_jobs_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/Job"
         },
         "title": "Response Jobs Api Jobs Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "summary": "Jobs"
   },
   "post": {
    "operationId": "start_api_jobs_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/Start"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Start Api Jobs Post",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Start"
   }
  },
  "/api/jobs/{job_id}": {
   "get": {
    "operationId": "one_job_api_jobs__job_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "job_id",
      "required": true,
      "schema": {
       "title": "Job Id",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Job"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "One Job"
   }
  },
  "/api/jobs/{job_id}/cancel": {
   "post": {
    "operationId": "cancel_api_jobs__job_id__cancel_post",
    "parameters": [
     {
      "in": "path",
      "name": "job_id",
      "required": true,
      "schema": {
       "title": "Job Id",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Cancel Api Jobs  Job Id  Cancel Post",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Cancel"
   }
  },
  "/api/jobs/{job_id}/events": {
   "get": {
    "operationId": "events_api_jobs__job_id__events_get",
    "parameters": [
     {
      "in": "path",
      "name": "job_id",
      "required": true,
      "schema": {
       "title": "Job Id",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Events"
   }
  },
...
 }
}